+++ packages/components/src/DataList/components/DataListLoadMore/DataListLoadMore.css.d.ts
@@ -1,5 +1,8 @@
 declare const styles: {
   readonly "trigger": string;
+  readonly "loadingMore": string;
+  readonly "backToTop": string;
+  readonly "fadeIn": string;
 };
 export = styles;